Update acc-netconf-5g to use ncclient rather than Python scripts
Previous versions of acc-netconf-5g used Python scripts included in the
Helm chart, but this caused issues when attempting to install the Python
libraries with a non-privileged user.

CU 3.2 and above provides the simple ncclient tool as part of the Docker
image which can be used instead to send the required snippets to the
server.

Co-authored-by: ilja <ilja.gucalo@accelleran.com>